create useModal hook for registration component
